Updated Dependencies to New Physic/Graphic Engine
Switched to PixiJS for rendering of graphics using WebGL. This should ensure
fast and efficient performance while guarunteeing easy future development. The
new physics engine in use is MatterJS. It is still updated and has collision
filtering, a necessary feature of a physics engine for this game, as balls
cannot collide with each other but they can collide with rectangles and walls.