Added Color as a Yarn/NPM Dependency
Color introduces the ability to calculate numeric values based on simple
RGB/HEX/HSV values. This feature is essential to working with PixiJS. Pixi
requires either hexadecimal or integer values of RGB. Given the intent to work
with HSL, this utility makes it easy for us to convert back and forth any time
in the future.