Added Application JS File, Constructor, Imports
Created the app.js file with the relevant constructor and class for launching a
default application. It launches the to be created Game class that will handle
the necessary game loading and playing components. It also imports all
dependencies including PEP, a polyfill for the Pointer Events API, the color
library for converting between color types, and PixiJS, the renderer for the
game we are creating.