Created Game Component Classes and Functions
We have created various game component classes for the game. Ball-Counter is
currently a placeholder for a text component that displays the number of
remaining balls to be shot.

Ball is the standard component for the launchable balls in the game. The ball
classes have the physics bodies and renderable components necessary for the ball
object in the game.

The rect class is representative of the static rectangles throughout the scene,
once again possesing the physics body and renderable component.

The collision-categories is a constants file holding the bitmasks/categories for
each type of collidable object. This allows for collision filtering between
balls and rectangles.

Finally, the launcher class possesses the necessary logic to render out the
launcher / targer based on user input and generates the necessary graphics.