Created Game Class and Game States
Created the Game Class and all the Game states for generating the gameplay and
handling the state machine. The Game States is equivalent to an enumeration
representing all the states the game could possibly be in. The Game class itself
instantiates all components of the game and manages its own state and the state
of other components it possesses. Altogether, the game class is what handles the
logic and responsiveness of the game.