Updated Launcher with Utility Functions
A new static utility function has been added to launcher to check the
minimum launch angle, as well as a constant for storing this minimum
angle. In addition, angles can now be converted beteween degrees and
radians using the appropriate toDegrees or toRadians functions from
util.js. In addition, a bug was fixed regarding ball.js and shall be
updated in rect.js, where the x and y coordinates of the shape didn't
properly correspond to the physics body.