Updated Dependencies and Architecture and Physics
Updated the dependencies to include matter-collisions plugin.
This enables for per body collision detection which is quite nice for
development, and reduces boilerplate code for checking for collisions.

There is a new architecture for the classes. We have started truly utilising the
prototypal inheritance and have added properties to the existing matter objects
rather than bundling the physics body and render body into a single class.

Basic physics now works for ball launching and all components are being
rendered. The majority of the logic has been put in place for the game.

We need to fix the position of the render body vs the physics body though.