Added Visible property and made balls bouncy
Added the visible property to the Rect and Balls created. This indicates to the
render loop that the component is renderable and has the getGraphic method,
rather than throwing a runtime error.

Also updated restitution to 1 for the ball, making it preserve velocity on
collision, preventing any stalling of the balls.